PyPI initial setup

+++ src/echaim/source_c/lib/ECHAIM.h
@@ -3,21 +3,21 @@
 
 #include "sqlite3.h"
 
-double * NmF2(double *lat, double *lon, double *year, double *month, double *day, \
-				double *hour, double *min, double *sec, int l0, int err);
+double * NmF2(double *lat, double *lon, int *year, int *month, int *day, \
+				int *hour, int *min, int *sec, int l0, int err);
 
-double * NmF2Storm(double *lat, double *lon, double *year, double *month, double *day, \
-				double *hour, double *min, double *sec, int l0, int err);
+double * NmF2Storm(double *lat, double *lon, int *year, int *month, int *day, \
+				int *hour, int *min, int *sec, int l0, int err);
 
-double * HmF2(double *lat, double *lon, double *year, double *month, double *day, \
-				double *hour, double *min, double *sec, int l0, int err);
+double * HmF2(double *lat, double *lon, int *year, int *month, int *day, \
+				int *hour, int *min, int *sec, int l0, int err);
 				
-double * HmF1(double *lat, double *lon, double *year, double *month, double *day, \
-				double *hour, double *min, double *sec, int l0, int err);
+double * HmF1(double *lat, double *lon, int *year, int *month, int *day, \
+				int *hour, int *min, int *sec, int l0, int err);
 
 double * densityPath(double *lat, double *lon, double *alt, \
-						double *year, double *month, double *day, \
-						double *hour, double *min, double *sec, int storm, int precip, int dregion, int l0, int err);
+						int *year, int *month, int *day, \
+						int *hour, int *min, int *sec, int storm, int precip, int dregion, int l0, int err);
 
 double ** densityProfile(double *lat, double *lon, int *year, int *month, int *day, \
 						int *hour, int *min, int *sec, int storm, int precip, int dregion, \
